+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * switch build tool to rollup
 * remove deprecated underscore names
+* fix revalidation for forms with novalidate attribute
 
 ## v0.10.2